cos/queue_service: extract drain stage into queue_service/drain.rs (#1035 P2)
Convert queue_service.rs → queue_service/mod.rs (directory module)
and extract the drain stage of the queue-service pipeline into a
sibling drain.rs:

- drain_exact_local_fifo_items_to_scratch (FIFO, ~95 LOC)
- drain_exact_local_items_to_scratch_flow_fair (MQFQ, ~120 LOC)
- drain_exact_prepared_fifo_items_to_scratch (FIFO prepared, ~105 LOC)
- drain_exact_prepared_items_to_scratch_flow_fair (MQFQ prepared, ~130 LOC)

The four fns implement the drain stage of the worker CoS service
pipeline: select → service → **drain** → release/restore → settle.
FIFO variants serve queues that don't need flow-fairness; the
_flow_fair variants implement MQFQ ordering across per-flow buckets
within a queue.

queue_service/mod.rs production LOC (excluding mod tests):
  2,999 → 2,558 (Phase 2; still over 2,000 threshold but moving —
  next phases would target select / service / settle stages)
queue_service/drain.rs: 461 LOC.

build clean (cargo build --release), 865 Rust tests pass. Cluster
smoke required (touches the worker drain hot path).
